requirements.txt
aiogram==3.10.0
aiohttp==3.9.5
pandas==2.2.2
openpyxl==3.1.3
python-docx==1.1.2
SQLAlchemy==2.0.30
asyncpg==0.29.0
APScheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1
